feat: реализовал блок "compilation" - в Postman проверил блок "Validation" и 75% блока "409 Conflict"

+++ ewm-service/src/main/java/ru/practicum/controller/admin/CategoryAdminController.java
@@ -1,6 +1,7 @@
 package ru.practicum.controller.admin;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -9,14 +10,20 @@ import ru.practicum.constraintGroup.Post;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.category.CategoryMapper;
 import ru.practicum.dto.category.NewCategoryDto;
+import ru.practicum.error.ConflictException;
 import ru.practicum.error.NotFoundException;
 import ru.practicum.model.Category;
 import ru.practicum.service.category.CategoryService;
+import ru.practicum.service.event.EventService;
+
+import javax.validation.constraints.NotNull;
 
 @RestController
 @RequiredArgsConstructor
+@Slf4j
 public class CategoryAdminController {
     private final CategoryService categoryService;
+    private final EventService eventService;
 
     @PostMapping("/admin/categories")
     @ResponseStatus(HttpStatus.CREATED)
@@ -29,14 +36,20 @@ public class CategoryAdminController {
 
     @DeleteMapping("/admin/categories/{catId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public boolean deleteCategories(@PathVariable Long catId) {
+    public boolean deleteCategories(@PathVariable @NotNull Long catId) {
         categoryService.getCategoryByIdOrElseThrow(catId);
 
-        return categoryService.deleteCategory(catId);
+        if (eventService.getEventsByCategoryId(catId).isEmpty()) {
+            return categoryService.deleteCategory(catId);
+        } else {
+            log.debug("Категория события с catId = {} используется событыями.", catId);
+            throw new ConflictException(String.format("Категория события с catId = %s используется событыями.",
+                    catId));
+        }
     }
 
     @PatchMapping("/admin/categories/{catId}")
-    public CategoryDto patchCategory(@PathVariable Long catId,
+    public CategoryDto patchCategory(@PathVariable @NotNull Long catId,
                                      @Validated({Patch.class}) @RequestBody NewCategoryDto newCategoryDto) throws NotFoundException {
         Category category = categoryService.getCategoryByIdOrElseThrow(catId);
         category.setName(newCategoryDto.getName());
